requirements.txt
fastapi
uvicorn
langchain
sentence-transformers
faiss-cpu
pypdf
beautifulsoup4
unstructured
python-dotenv
tqdm
bytez
rank-bm25
python-multipart
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
slack_sdk
notion-client
apscheduler